refactor: RacingCars에 일급 컬렉션 적용
- 기존의 InputValidator의 검증 기능을 RacingCars로 이동
- 일급 컬렉션 적용
